backend/pyproject.toml
[project]
name = "raffle-backend"
version = "1.0.0"
description = "Backend API for ticket raffle system"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "pydantic>=2.5.0",
    "pydantic[email]>=2.5.0",
    "python-multipart>=0.0.6",
    "instaloader>=4.10.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "httpx>=0.25.0"]
